fix: Reduce the frequency of updates when there are no windows
This change drops the frequency of background updates from 1 minute to 5 minutes when there are no windows. This ensures content is relatively up-to-date when foregrounding the application but means we're not doing too much work in the background.

This also introduces a new scene model in preparation for presenting errors and alerts to the user.