fix: Disable multiple scenes on iPadOS (#227)
This change disables multiple scenes on iPadOS to address issues where
multiple scenes were being created when opening the app from the App
Library (check out
https://github.com/jbmorley/swiftui-multi-window-issues/ for a more
involved discussion on the issues seen). Ultimately this is a
simplifying decision--the app doesn't feel like it gains a lot from
having multiple windows and they seem precarious enough in SwiftUI that
it's just not worth the development time and resulting inconsistent user
experience. We can revisit this in the future if the APIs improve and
users really want this functionality.