fix: Unable to sign out on iPad (#228)
This change addresses the issue where the iPad wasn't displaying the
sign out confirmation. It switches to use an alert instead of SwiftUI's
built in confirmation dialogs as these present as a popover on iPadOS
which doesn't match user expectations in this use-case. Disappointingly,
the change also includes a fairly involved approach designed to track
the presented sheets to allow us to guarantee the confirmation is always
presented from the top-most sheet on all platforms.