fix: Use up-to-date GitHub authentication details from iCloud keychain (#267)
This change ensures clients always use the most up-to-date GitHub
authentication details from iCloud keychain. Prior to this change it was
necessary to relaunch Builds to read the latest settings from iCloud.

Under the hood, this changes to a transient GitHub client instance,
reading the authentication token on every new fetch.