refactor: Rename `WorkflowIdentifier` to `WorkflowIdentifierEntity` (#338)
It seems like the magic intent synthesis doesn't work if the intents (or
their structs) are defined outside of the target that uses them. This
means, for the time being at least, we're stuck with the duplicate
struct in the widget target. This change renames that struct
(`WorkflowIdentifier`) to `WorkflowIdentifierEntity` to explicitly
acknowledge this and makes it a simple wrapper for
`WorkflowInstance.ID`.