fix: Show workflow names in the Single Workflow widget settings (#375)
This change updates `WorkflowIdentifier` to store a cached snapshot of
the workflow name at the time the identifier was created. This ensures
we always have _something_ to display when looking up the workflow name
and prevents us from falling back on the workflow id (which makes no
sense to users). Long term, it would be good to store a local cache of
workflow names that we can use to provide an updated name.